silently check mvn existence on startup
