+++ src/archetype/ArchetypeModule.ts
@@ -5,8 +5,8 @@ import * as fse from "fs-extra";
 import * as path from "path";
 import { QuickPickItem, Uri, window, workspace } from "vscode";
 import { instrumentOperationStep, sendInfo } from "vscode-extension-telemetry-wrapper";
-import { OperationCanceledError } from "../Errors";
 import { getMavenLocalRepository, getPathToExtensionRoot } from "../utils/contextUtils";
+import { OperationCanceledError } from "../utils/errorUtils";
 import { executeInTerminal, getEmbeddedMavenWrapper, getMaven } from "../utils/mavenUtils";
 import { openDialogForFolder } from "../utils/uiUtils";
 import { Utils } from "../utils/Utils";
